requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
python-dotenv>=1.0.0
supabase>=1.0.3
crewai>=0.1.33
langchain-openai>=0.0.2
google-cloud-storage>=2.10.0
requests>=2.31.0
PyMuPDF>=1.23.5